Revert "Fixing poll server tests, moving them to fides client"
This reverts commit bbbe426b37b16bb72ba834989ffdc4304574a320.
